+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -152,11 +152,12 @@
     <Compile Include="Routing\DistributionStrategyScope.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\AssemblyPublisherSource.cs" />
     <Compile Include="Routing\AssemblyRouteSource.cs" />
-    <Compile Include="Routing\ConfiguredPublishers.cs" />
+    <Compile Include="Routing\MessageDrivenSubscriptions\ConfiguredPublishers.cs" />
     <Compile Include="Routing\ConfiguredUnicastRoutes.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\IPublisherSource.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\NamespacePublisherSource.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\TypePublisherSource.cs" />
+    <Compile Include="Routing\NativePublishSubscribeFeature.cs" />
     <Compile Include="Routing\TypeRouteSource.cs" />
     <Compile Include="Routing\IDistributionPolicy.cs" />
     <Compile Include="Routing\IRouteSource.cs" />
